deploy/requirements.txt
virtualenv
aws-cdk.aws-s3
aws-cdk.aws-lambda
aws-cdk.assets
aws-cdk.aws_ec2
aws_cdk.aws_apigateway
aws_cdk.aws_cloudfront
aws-cdk.aws_s3_deployment
aws-cdk.aws_cognito
aws-cdk.aws_eks
aws-cdk.aws_ecr
aws-cdk.aws_ecr_assets
aws_cdk.aws_directoryservice
aws_cdk.aws-rds